requirements.txt
pandas
numpy
pyyaml
scikit-learn
joblib
mlflow
dvc
lightgbm
statsmodels
prophet
tensorflow>=2.10
uvicorn[standard]
fastapi
python-multipart
matplotlib
evidently
prometheus-client